lesson-3.5

+++ src/main/java/ru/romankuznetsov/dto/OrderItemDto.java
@@ -4,12 +4,14 @@ import lombok.Data;
 import ru.romankuznetsov.entity.OrderItem;
 import ru.romankuznetsov.entity.Product;
 
+import java.math.BigDecimal;
+
 @Data
 public class OrderItemDto {
     private Long productId;
     private String productTitle;
-    private Integer pricePerProduct;
-    private Integer price;
+    private BigDecimal pricePerProduct;
+    private BigDecimal price;
     private Integer quantity;
 
     public OrderItemDto(OrderItem orderItem) {
